Exclude backlinking to self
Since Joplin 2.13.4, the search API has been changed to search using
the note's ID (returning the original note), if a text content search is
unsuccessful. This has the effect of including a backlink to the original
note if no notes link to it.

Filters the search results returned to not link to the original note, as
determined by the note's ID.

Closes #17